scripts/requirements.txt
pandas
pyarrow
openpyxl
folium
streamlit-folium
plotly
streamlit
watchdog
ollama
pinecone-client>=3.0.0
pypdf
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-ollama
langchain-pinecone
langgraph
langchain-core
duckdb
scikit-learn
matplotlib
seaborn
requests
fpdf
# Phase 1 — forecasting
prophet
statsmodels
# Phase 3 — UI
streamlit-lottie
# Phase 4 — observability
prometheus-client
pytest
